extract runtime pallets part.2
