+++ runtime/crab/src/pallets/mod.rs
@@ -6,3 +6,9 @@ pub use babe::*;
 
 pub mod timestamp;
 pub use timestamp::*;
+
+pub mod balances;
+pub use balances::*;
+
+pub mod transaction_payment;
+pub use transaction_payment::*;
